Moved Debriefing endpoint to a mixin in the debriefings model
